steam window and harvesters update
This implements showing of the buffer stream (makes it a little slower) and an update to the harvesters. Unfortunately the harvesters are still just the same as the cv2 pipeline, just more verbose. The thing that we are looking for is to be able to do a "binning" on the pixel ON THE DEVICE and speed up the acquision frame rate