homing switch goosechase and fluor update
These were a series of changes that I mad while trying to figure out why i was only getting an alarm from the blackboxes. Turns out that if you enable the hard limit switches then the system will need to home BEFORE anything else. This includes the $X unlock command for some reason. I have no idea why

Additionally I added a setting in the plate parameters that allows for the choosing of a specific fluorescent excitation profile (UBGR and 0-100%), default is that is everything is always on and at full tilt