autofocus grid fit
this is an attempt for when you add a plate for the first time then the system will go through a more complete autofocus (completely blind) to all the wells. After that then the z heights are found and a quick autofocus is ran on the plates just to make sure its not too far off.
this is stored as a txt file on the output folder